Add files via upload
My new project  name  is  a Random meme generator.  It  can  produce  memes  randomly. I made this project by using my skills of HTML,CSS and Javascript. The images which are shown in this project are taken from the "MEMES API"... It was a phenomenal experience i had while doing this project. The Random meme generator project aims to address the need for a user-friendly and efficient tool for creating memes with minimal effort. The challenge lies in seamlessly integrating text input , image selection, and  Random meme generation to provide a quick and enjoyable user experience...

API Reliability: The Random Meme Generator heavily relies on the img flip API for image retrieval. Any disruptions or unreliability in the API service could impact the operational efficiency of the generator. Continuous monitoring and a contingency 

A meme is a typically humorous image, video, or text that conveys a particular phenomenon or theme and is spread among individuals.Among the different types of memes, image-based memes are the most popular. They consist of an image (a template),short text, and humor, making them efﬁcient in delivering information to the audience. Compared to other sources of information like books, articles, videos, television, and radio, image memes require minimal immersion in the content and spread virally across the Internet and between cultures.